VFD rework (#1324)
* Fixed Generate_VCXProj

* Moved VFD spindles to their own namespace and inheritance structure. TODO: Configuration.

* Changed the factory a bit to ensure configurations still work.

* This compiles. TODO: Loader script.

* Changed names and made them consistent. This also removes them from IRAM.

* Split out the VFDProtocol code into a separate file.

* Added comment

---------
