+++ FluidNC/src/Platform.h
@@ -9,5 +9,6 @@
 #else
 
 #    define WEAK_LINK
+#    define IRAM_ATTR
 
 #endif
