+++ FluidNC/src/Spindles/VFDSpindle.h
@@ -14,61 +14,28 @@
 
 namespace Spindles {
     extern Uart _uart;
+    
+    
+    namespace VFD {
+        // VFDProtocol resides in a separate class because it doesn't need to be in IRAM. This contains all the
+        // VFD specific code, which is called from a separate task.
+        class VFDProtocol;
+    }
 
-    class VFD : public Spindle {
+    // VFD base class. Called by the stepper engine. Normally you don't want to touch this.
+    class VFDSpindle : public Spindle {
     private:
-        static const int VFD_RS485_MAX_MSG_SIZE = 16;  // more than enough for a modbus message
-        static const int MAX_RETRIES            = 5;   // otherwise the spindle is marked 'unresponsive'
+        friend class Spindles::VFD::VFDProtocol;
 
-        void set_mode(SpindleState mode, bool critical);
+        VFD::VFDProtocol* detail_ = nullptr;
 
         int32_t  _current_dev_speed   = -1;
         uint32_t _last_speed          = 0;
         Percent  _last_override_value = 100;  // no override is 100 percent
 
-        static QueueHandle_t vfd_cmd_queue;
-        static TaskHandle_t  vfd_cmdTaskHandle;
-        static void          vfd_cmd_task(void* pvParameters);
-
-        static uint16_t ModRTU_CRC(uint8_t* buf, int msg_len);
-        enum VFDactionType : uint8_t { actionSetSpeed, actionSetMode };
-        struct VFDaction {
-            VFDactionType action;
-            bool          critical;
-            uint32_t      arg;
-        };
+        void set_mode(SpindleState mode, bool critical);
 
     protected:
-        struct ModbusCommand {
-            bool critical;  // TODO SdB: change into `uint8_t critical : 1;`: We want more flags...
-
-            uint8_t tx_length;
-            uint8_t rx_length;
-            uint8_t msg[VFD_RS485_MAX_MSG_SIZE];
-        };
-
-    private:
-        bool prepareSetModeCommand(SpindleState mode, ModbusCommand& data);
-        bool prepareSetSpeedCommand(uint32_t speed, ModbusCommand& data);
-
-        static void reportParsingErrors(ModbusCommand cmd, uint8_t* rx_message, size_t read_length);
-        static void reportCmdErrors(ModbusCommand cmd, uint8_t* rx_message, size_t read_length, uint8_t id);
-
-    protected:
-        // Commands:
-        virtual void direction_command(SpindleState mode, ModbusCommand& data) = 0;
-        virtual void set_speed_command(uint32_t rpm, ModbusCommand& data)      = 0;
-
-        // Commands that return the status. Returns nullptr if unavailable by this VFD (default):
-        using response_parser = bool (*)(const uint8_t* response, VFD* spindle);
-
-        virtual response_parser initialization_sequence(int index, ModbusCommand& data) { return nullptr; }
-        virtual response_parser get_current_speed(ModbusCommand& data) { return nullptr; }
-        virtual response_parser get_current_direction(ModbusCommand& data) { return nullptr; }
-        virtual response_parser get_status_ok(ModbusCommand& data) = 0;
-        virtual bool            safety_polling() const { return true; }
-        bool                    use_delay_settings() const override { return true; }
-
         // The constructor sets these
         int     _uart_num  = -1;
         Uart*   _uart      = nullptr;
@@ -79,11 +46,11 @@ namespace Spindles {
         volatile bool _syncing;
 
     public:
-        VFD(const char* name) : Spindle(name) {}
-        VFD(const VFD&)            = delete;
-        VFD(VFD&&)                 = delete;
-        VFD& operator=(const VFD&) = delete;
-        VFD& operator=(VFD&&)      = delete;
+        VFDSpindle(const char* name, VFD::VFDProtocol* detail) : Spindle(name), detail_(detail) {}
+        VFDSpindle(const VFDSpindle&) = delete;
+        VFDSpindle(VFDSpindle&&)      = delete;
+        VFDSpindle& operator=(const VFDSpindle&) = delete;
+        VFDSpindle& operator=(VFDSpindle&&) = delete;
 
         void init();
         void config_message();
@@ -97,6 +64,6 @@ namespace Spindles {
         void validate() override;
         void group(Configuration::HandlerBase& handler) override;
 
-        virtual ~VFD() {}
+        virtual ~VFDSpindle() {}
     };
 }
